+++ AdventOfCode/2015/DayTen.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using AdventOfCode.Core;
 
 namespace AdventOfCode._2015
 {
@@ -19,7 +20,7 @@ namespace AdventOfCode._2015
             _cache = new Dictionary<string, string>();
         }
 
-        public int SolvePart1()
+        public long SolvePart1()
         {
             return SolvePart1_Str().Length;
         }
@@ -27,7 +28,7 @@ namespace AdventOfCode._2015
         public string SolvePart1_Str()
         {
             var morph = _start;
-            for(var round=0;round<_rounds;round++)
+            for (var round = 0; round < _rounds; round++)
             {
                 morph = Expand(morph);
             }
@@ -50,28 +51,28 @@ namespace AdventOfCode._2015
             {
                 var current = morph[0];
                 var count = 1;
-                for(var cIdx=1;cIdx<morph.Length;cIdx++)
+                for (var cIdx = 1; cIdx < morph.Length; cIdx++)
                 {
                     if (morph[cIdx] == current) count++;
                     else
                     {
                         sb.Append(count);
                         sb.Append(current);
-                        
+
                         count = 1;
                         current = morph[cIdx];
                     }
                 }
                 sb.Append(count);
                 sb.Append(current);
-                
+
             }
 
             _cache.Add(morph, sb.ToString());
             return sb.ToString();
         }
 
-        public int SolvePart2()
+        public long SolvePart2()
         {
             throw new NotImplementedException();
         }
